refact scrollNodeToCenter
