+++ src/canva/constants.ts
@@ -6,3 +6,8 @@ export const ZOOM_KEY_CODES = {
   FIT_TO_VIEW: 48,
   FIT_TO_HUNDRED: 49,
 };
+
+export const isSafari = /^((?!chrome|android).)*safari/i.test(
+  navigator.userAgent
+);
+export const TIME_TO_WAIT = isSafari ? 600 : 300;
